Seperate keystore specific functions in util and apply deciperer
